separated chorelist class into a separate file
