+++ Chorelord/ChoreListViewController.swift
@@ -14,9 +14,11 @@ let employeeNames = ["Baylee", "Alex", "Anya", "Ellie", "Lee", "Dean", "Tony", "
 
 class ChoreListViewController: UITableViewController {
   
+  var choreList = ChoreList(choreNames: choreNames, employeeNames: employeeNames)
+  
   override func viewDidLoad() {
     super.viewDidLoad()
-    reorderList(employeeNames, start: "2016-02-01", end: "2016-02-15", interval: 7) // this is doing nothing right now, just for testing purposes
+    choreList.reorderList(employeeNames, start: "2016-02-01", end: "2016-02-15", interval: 7) // this is doing nothing right now, just for testing purposes
     tableView.registerNib(UINib(nibName: "ChoreTableViewCell", bundle: nil), forCellReuseIdentifier: "ChoreTableViewCell")
   }
   
@@ -31,39 +33,4 @@ class ChoreListViewController: UITableViewController {
     return cell
   }
   
-  /**
-  * Format string date into NSDate
-  **/
-  func formatDate(date: String) -> NSDate {
-    let dateFormatter = NSDateFormatter()
-    dateFormatter.dateFormat = "yyyy-MM-dd" // ex: 2015-02-01
-    let formatted: NSDate = dateFormatter.dateFromString(date)!
-    return formatted
   }
-  
-  /**
-  * Calculate the day difference between 2 dates
-  **/
-  func calcDayDiff(start: String, end: String) -> Int {
-    let calendar = NSCalendar.currentCalendar()
-    let unit: NSCalendarUnit = NSCalendarUnit.Day
-    let components = calendar.components(unit, fromDate: self.formatDate(start), toDate: self.formatDate(end), options: NSCalendarOptions(rawValue: 0))
-    return components.day
-  }
-  
-  /**
-  * Reorder the list based on the cycle
-  **/
-  func reorderList(nameList: [String], start: String, end: String, interval: Int) -> [String] {
-    var newList = nameList
-    let dayDifference = calcDayDiff(start, end: end)
-    let cycleNumber = dayDifference/interval // always floor
-    if (cycleNumber > 0) {
-      for _ in 1...cycleNumber {
-        let first = newList.removeFirst()
-        newList.append(first)
-      }
-    }
-    return newList
-  }
-}
